Updated router files for more functionality
Updated router files for more functionality on the website. Allowed for user to join clean ups, added a user_id field for when creating a clean up, added the ability to get all the clean ups data based on the user currently logged in and also allowed for them to delete and edit the clean ups as well.